pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["Vadim Sultanov <vadim.soultanov@mail.ru>"]
readme = "README.md"
packages = [{include = "fastapi_template"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["session"], version = "^0.110.0"}
sqlmodel = "^0.0.16"
alembic = "^1.13.1"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
bcrypt = "^4.1.2"
jwt = "^1.3.1"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^2.6.4"}
pydantic-settings = "^2.2.1"
dramatiq = {extras = ["watch"], version = "^1.16.0"}
pika = "^1.3.2"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
pre-commit = "^3.7.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
mimesis = "^15.1.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
